Working version: data in Excel, work in Python
The architecture changed from using a SQLite database to using an Excel
document on my computer.  This was easier to manage than a database, and
I anticipate that either way the database would need to be re-written if
this migrated to a Cloud application.  The free databases tend to be
NoSQL, but those often aren't runnable locally.

This working version was used to send out the first round of lunch
roulette emails.