Implement an announcement feature
This change refactors the Powershell script to take a hashtable of
replacements, and to validate that all expected replacements are
completed before an email is sent.  This flexibility is helpful
because announcements won't have as many variables to replace.
But we want to be sure that no variables are accidentally
lingering.